# Decision log and QueryTracer refactors