--- a/plugins/logs/plugin_benchmark_test.go
+++ b/plugins/logs/plugin_benchmark_test.go
@@ -2,6 +2,7 @@ package logs
 
 import (
 	"context"
+	"fmt"
 	"testing"
 
 	"github.com/open-policy-agent/opa/plugins"
@@ -169,6 +170,44 @@ func BenchmarkMaskingNop(b *testing.B) {
 		}
 	}
 
+}
+
+func BenchmarkMaskingRuleCountsNop(b *testing.B) {
+	numRules := []int{1, 10, 100, 1000}
+
+	ctx := context.Background()
+	store := inmem.New()
+
+	manager, err := plugins.New(nil, "test", store)
+	if err != nil {
+		b.Fatal(err)
+	} else if err := manager.Start(ctx); err != nil {
+		b.Fatal(err)
+	}
+
+	cfg := &Config{Service: "svc"}
+	cfg.validateAndInjectDefaults([]string{"svc"}, nil)
+	plugin := New(cfg, manager)
+
+	for _, ruleCount := range numRules {
+
+		b.Run(fmt.Sprintf("%dRules", ruleCount), func(b *testing.B) {
+			b.ResetTimer()
+			for i := 0; i < b.N; i++ {
+				b.StopTimer()
+				var event EventV1
+				if err := util.UnmarshalJSON([]byte(largeEvent), &event); err != nil {
+					b.Fatal(err)
+				}
+
+				b.StartTimer()
+
+				if err := plugin.maskEvent(ctx, nil, &event); err != nil {
+					b.Fatal(err)
+				}
+			}
+		})
+	}
 }
 
 func BenchmarkMaskingErase(b *testing.B) {