--- a/server/buffer.go
+++ b/server/buffer.go
@@ -7,6 +7,7 @@ package server
 import (
 	"time"
 
+	"github.com/open-policy-agent/opa/ast"
 	"github.com/open-policy-agent/opa/metrics"
 	"github.com/open-policy-agent/opa/storage"
 	"github.com/open-policy-agent/opa/topdown"
@@ -34,6 +35,7 @@ type Info struct {
 	Path       string
 	Timestamp  time.Time
 	Input      *interface{}
+	InputAST   ast.Value
 	Results    *interface{}
 	Error      error
 	Metrics    metrics.Metrics