… CHANGELOG.md
… plugins/logs/plugin_test.go
… profiler/profiler_bench_test.go
… profiler/profiler_test.go
… rego/example_test.go
… rego/rego_test.go
… test/authz/authz_bench_test.go
… test/e2e/authz/authz_bench_integration_test.go
… test/e2e/concurrency/concurrency_test.go
… test/e2e/diagnostics/diagnostics_test.go
… test/e2e/logs/console/console_decision_logger_benchmark_test.go
… test/e2e/logs/console/console_decision_logger_test.go
… test/e2e/logs/remote/remote_decision_logger_benchmark_test.go
… test/e2e/logs/utils.go
… test/e2e/metrics/metrics_test.go
… test/e2e/testing.go
… tester/runner_test.go
… topdown/topdown_bench_test.go