There are a handful of changes grouped together, split out from #2483, which are the performance enhancements and refactor commits.

At a high level this is plumbing the new QueryTracer through the Rego package, improving the performance for decision logging, and refactoring tests to better benchmark the decision logging.

See individual commit messages for more details