@patrick-east 2020-06-30
Benchmarks for this change compared to current (as of right now) master:
```
{16:59} ~/p/g/s/g/o/opa:issue/2089-refactors-and-perf ◈ ❯ benchstat before.txt after.txt
name                                              old time/op    new time/op    delta
pkg:github.com/open-policy-agent/opa/test/e2e/authz goos:linux goarch:amd64
RESTAuthzForbidAuthn-24                              184µs ± 2%     185µs ± 1%     ~     (p=0.762 n=10+8)
RESTAuthzForbidPath-24                               250µs ± 1%     250µs ± 2%     ~     (p=1.000 n=10+10)
RESTAuthzForbidMethod-24                             254µs ± 1%     253µs ± 1%     ~     (p=0.497 n=10+9)
RESTAuthzAllow10Paths-24                             254µs ± 1%     253µs ± 1%     ~     (p=0.315 n=10+10)
RESTAuthzAllow100Paths-24                            679µs ± 3%     661µs ± 4%   -2.70%  (p=0.009 n=10+10)
RESTAuthzAllow1000Paths-24                          4.42ms ± 2%    4.26ms ± 2%   -3.55%  (p=0.000 n=10+10)
pkg:github.com/open-policy-agent/opa/test/e2e/logs/console goos:linux goarch:amd64
RESTConsoleDecisionLogger/1x0-24                     291µs ± 1%     264µs ± 1%   -9.25%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/10x0-24                    295µs ± 1%     269µs ± 1%   -8.84%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/100x0-24                   299µs ± 1%     274µs ± 1%   -8.35%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/1000x0-24                  296µs ± 1%     271µs ± 1%   -8.22%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/1x1-24                     310µs ± 0%     292µs ± 1%   -5.77%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/10x1-24                    310µs ± 1%     293µs ± 1%   -5.75%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/100x1-24                   315µs ± 1%     296µs ± 1%   -6.18%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/1000x1-24                  312µs ± 1%     295µs ± 1%   -5.59%  (p=0.000 n=10+9)
RESTConsoleDecisionLogger/10x10-24                   334µs ± 0%     314µs ± 1%   -6.02%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/100x10-24                  332µs ± 1%     315µs ± 1%   -5.20%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/1000x10-24                 334µs ± 0%     313µs ± 0%   -6.46%  (p=0.000 n=10+8)
RESTConsoleDecisionLogger/100x100-24                 493µs ± 2%     487µs ± 1%   -1.12%  (p=0.004 n=10+10)
RESTConsoleDecisionLogger/1000x100-24                496µs ± 0%     468µs ± 0%   -5.62%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/1000x1000-24              2.01ms ± 1%    1.96ms ± 1%   -2.28%  (p=0.000 n=10+10)
pkg:github.com/open-policy-agent/opa/test/e2e/logs/remote goos:linux goarch:amd64
RESTDecisionLogAuthzForbidAuthn-24                   273µs ± 3%     243µs ± 2%  -11.16%  (p=0.000 n=10+10)
RESTDecisionLogAuthzForbidPath-24                    330µs ± 1%     308µs ± 1%   -6.68%  (p=0.000 n=10+9)
RESTDecisionLogAuthzForbidMethod-24                  336µs ± 2%     312µs ± 1%   -7.30%  (p=0.000 n=10+10)
RESTDecisionLogAuthzAllow10Paths-24                  333µs ± 1%     308µs ± 2%   -7.50%  (p=0.000 n=9+10)
RESTDecisionLogAuthzAllow100Paths-24                 736µs ± 3%     728µs ± 2%     ~     (p=0.105 n=10+10)
RESTDecisionLogAuthzAllow1000Paths-24               4.35ms ± 2%    4.37ms ± 2%     ~     (p=0.684 n=10+10)
RESTRemoteDecisionLogger/1x0-24                      235µs ± 3%     222µs ± 3%   -5.55%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/10x0-24                     237µs ± 2%     220µs ± 4%   -7.34%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/100x0-24                    239µs ± 5%     220µs ± 3%   -7.72%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/1000x0-24                   240µs ± 3%     220µs ± 2%   -8.07%  (p=0.000 n=9+10)
RESTRemoteDecisionLogger/1x1-24                      249µs ± 2%     233µs ± 3%   -6.61%  (p=0.000 n=9+10)
RESTRemoteDecisionLogger/10x1-24                     246µs ± 3%     231µs ± 3%   -5.81%  (p=0.000 n=10+9)
RESTRemoteDecisionLogger/100x1-24                    246µs ± 2%     234µs ± 4%   -4.50%  (p=0.000 n=9+10)
RESTRemoteDecisionLogger/1000x1-24                   245µs ± 4%     232µs ± 4%   -5.25%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/10x10-24                    267µs ± 2%     244µs ± 2%   -8.72%  (p=0.000 n=9+10)
RESTRemoteDecisionLogger/100x10-24                   270µs ± 3%     241µs ± 5%  -10.50%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/1000x10-24                  268µs ± 2%     246µs ± 3%   -8.41%  (p=0.000 n=8+9)
RESTRemoteDecisionLogger/100x100-24                  368µs ± 2%     368µs ± 4%     ~     (p=0.971 n=10+10)
RESTRemoteDecisionLogger/1000x100-24                 370µs ± 2%     369µs ± 2%     ~     (p=1.000 n=10+10)
RESTRemoteDecisionLogger/1000x1000-24               1.47ms ± 1%    1.49ms ± 1%   +1.26%  (p=0.004 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1x0-24           341µs ± 2%     315µs ± 2%   -7.58%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/10x0-24          342µs ± 1%     317µs ± 2%   -7.48%  (p=0.000 n=8+10)
RESTRemoteDecisionLoggerMaskApplied/100x0-24         338µs ± 2%     316µs ± 2%   -6.60%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1000x0-24        338µs ± 3%     320µs ± 2%   -5.42%  (p=0.000 n=10+9)
RESTRemoteDecisionLoggerMaskApplied/1x1-24           347µs ± 2%     329µs ± 2%   -5.22%  (p=0.000 n=10+9)
RESTRemoteDecisionLoggerMaskApplied/10x1-24          346µs ± 1%     330µs ± 3%   -4.48%  (p=0.000 n=9+10)
RESTRemoteDecisionLoggerMaskApplied/100x1-24         346µs ± 2%     328µs ± 3%   -5.19%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1000x1-24        347µs ± 2%     328µs ± 2%   -5.73%  (p=0.000 n=10+9)
RESTRemoteDecisionLoggerMaskApplied/10x10-24         358µs ± 3%     341µs ± 2%   -4.79%  (p=0.000 n=10+9)
RESTRemoteDecisionLoggerMaskApplied/100x10-24        356µs ± 4%     340µs ± 4%   -4.31%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1000x10-24       356µs ± 3%     339µs ± 2%   -4.70%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/100x100-24       476µs ± 3%     462µs ± 1%   -3.02%  (p=0.000 n=10+8)
RESTRemoteDecisionLoggerMaskApplied/1000x100-24      474µs ± 5%     459µs ± 3%   -3.18%  (p=0.011 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1000x1000-24    1.59ms ± 2%    1.60ms ± 1%     ~     (p=0.829 n=10+8)

name                                              old alloc/op   new alloc/op   delta
pkg:github.com/open-policy-agent/opa/test/e2e/authz goos:linux goarch:amd64
RESTAuthzForbidAuthn-24                             30.3kB ± 0%    30.3kB ± 0%     ~     (p=0.483 n=10+9)
RESTAuthzForbidPath-24                              46.4kB ± 0%    46.4kB ± 0%     ~     (p=0.906 n=7+10)
RESTAuthzForbidMethod-24                            47.0kB ± 0%    47.0kB ± 0%     ~     (p=0.945 n=10+8)
RESTAuthzAllow10Paths-24                            45.7kB ± 0%    45.8kB ± 0%     ~     (p=0.184 n=10+10)
RESTAuthzAllow100Paths-24                            176kB ± 0%     175kB ± 0%     ~     (p=0.182 n=10+9)
RESTAuthzAllow1000Paths-24                          1.47MB ± 0%    1.47MB ± 0%     ~     (p=0.579 n=10+10)
pkg:github.com/open-policy-agent/opa/test/e2e/logs/console goos:linux goarch:amd64
RESTConsoleDecisionLogger/1x0-24                    53.8kB ± 1%    47.5kB ± 1%  -11.78%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/10x0-24                   54.1kB ± 1%    47.6kB ± 0%  -12.07%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/100x0-24                  54.4kB ± 1%    48.0kB ± 1%  -11.79%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/1000x0-24                 52.8kB ± 0%    46.3kB ± 0%  -12.32%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/1x1-24                    57.8kB ± 1%    53.7kB ± 1%   -7.12%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/10x1-24                   57.9kB ± 0%    53.6kB ± 0%   -7.53%  (p=0.000 n=9+8)
RESTConsoleDecisionLogger/100x1-24                  58.1kB ± 1%    53.7kB ± 0%   -7.56%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/1000x1-24                 56.1kB ± 0%    51.9kB ± 0%   -7.36%  (p=0.000 n=10+9)
RESTConsoleDecisionLogger/10x10-24                  65.5kB ± 0%    61.2kB ± 1%   -6.56%  (p=0.000 n=9+10)
RESTConsoleDecisionLogger/100x10-24                 65.5kB ± 0%    61.4kB ± 0%   -6.29%  (p=0.000 n=10+9)
RESTConsoleDecisionLogger/1000x10-24                63.4kB ± 0%    59.3kB ± 0%   -6.47%  (p=0.000 n=10+9)
RESTConsoleDecisionLogger/100x100-24                 144kB ± 0%     140kB ± 0%   -2.72%  (p=0.000 n=9+10)
RESTConsoleDecisionLogger/1000x100-24                140kB ± 0%     136kB ± 0%   -3.05%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/1000x1000-24               970kB ± 0%     968kB ± 0%   -0.25%  (p=0.000 n=10+9)
pkg:github.com/open-policy-agent/opa/test/e2e/logs/remote goos:linux goarch:amd64
RESTDecisionLogAuthzForbidAuthn-24                  58.0kB ± 0%    53.7kB ± 0%   -7.33%  (p=0.000 n=9+10)
RESTDecisionLogAuthzForbidPath-24                   73.7kB ± 1%    69.5kB ± 0%   -5.77%  (p=0.000 n=10+10)
RESTDecisionLogAuthzForbidMethod-24                 74.3kB ± 0%    70.3kB ± 0%   -5.48%  (p=0.000 n=9+10)
RESTDecisionLogAuthzAllow10Paths-24                 72.9kB ± 0%    68.8kB ± 0%   -5.61%  (p=0.000 n=9+10)
RESTDecisionLogAuthzAllow100Paths-24                 203kB ± 0%     199kB ± 0%   -2.10%  (p=0.000 n=10+10)
RESTDecisionLogAuthzAllow1000Paths-24               1.50MB ± 0%    1.49MB ± 0%   -0.26%  (p=0.000 n=8+9)
RESTRemoteDecisionLogger/1x0-24                     52.3kB ± 0%    45.7kB ± 0%  -12.50%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/10x0-24                    52.2kB ± 0%    45.8kB ± 0%  -12.22%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/100x0-24                   52.2kB ± 0%    45.8kB ± 0%  -12.31%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/1000x0-24                  52.3kB ± 0%    45.8kB ± 0%  -12.44%  (p=0.000 n=9+10)
RESTRemoteDecisionLogger/1x1-24                     55.6kB ± 0%    51.5kB ± 0%   -7.29%  (p=0.000 n=10+9)
RESTRemoteDecisionLogger/10x1-24                    55.5kB ± 0%    51.5kB ± 0%   -7.26%  (p=0.000 n=9+9)
RESTRemoteDecisionLogger/100x1-24                   55.5kB ± 0%    51.5kB ± 0%   -7.26%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/1000x1-24                  55.5kB ± 0%    51.5kB ± 0%   -7.17%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/10x10-24                   62.7kB ± 0%    58.6kB ± 0%   -6.43%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/100x10-24                  62.7kB ± 0%    58.6kB ± 0%   -6.58%  (p=0.000 n=10+8)
RESTRemoteDecisionLogger/1000x10-24                 62.6kB ± 0%    58.7kB ± 0%   -6.31%  (p=0.000 n=10+9)
RESTRemoteDecisionLogger/100x100-24                  138kB ± 0%     134kB ± 0%   -2.91%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/1000x100-24                 138kB ± 0%     134kB ± 0%   -2.87%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/1000x1000-24                943kB ± 0%     939kB ± 0%   -0.41%  (p=0.000 n=9+8)
RESTRemoteDecisionLoggerMaskApplied/1x0-24          96.1kB ± 0%    89.6kB ± 0%   -6.71%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/10x0-24         96.0kB ± 0%    89.6kB ± 0%   -6.74%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/100x0-24        96.1kB ± 0%    89.6kB ± 0%   -6.79%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1000x0-24       96.1kB ± 0%    89.5kB ± 0%   -6.84%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1x1-24          99.3kB ± 0%    95.3kB ± 0%   -4.01%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/10x1-24         99.2kB ± 0%    95.3kB ± 0%   -3.97%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/100x1-24        99.3kB ± 0%    95.4kB ± 0%   -3.94%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1000x1-24       99.3kB ± 0%    95.3kB ± 0%   -4.09%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/10x10-24         106kB ± 0%     102kB ± 0%   -3.79%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/100x10-24        106kB ± 0%     102kB ± 0%   -3.71%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1000x10-24       106kB ± 0%     103kB ± 0%   -3.70%  (p=0.000 n=9+10)
RESTRemoteDecisionLoggerMaskApplied/100x100-24       181kB ± 0%     177kB ± 0%   -2.23%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1000x100-24      181kB ± 0%     177kB ± 0%   -2.15%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1000x1000-24     987kB ± 0%     983kB ± 0%   -0.41%  (p=0.000 n=10+9)

name                                              old allocs/op  new allocs/op  delta
pkg:github.com/open-policy-agent/opa/test/e2e/authz goos:linux goarch:amd64
RESTAuthzForbidAuthn-24                                409 ± 0%       409 ± 0%     ~     (all equal)
RESTAuthzForbidPath-24                                 828 ± 0%       828 ± 0%     ~     (all equal)
RESTAuthzForbidMethod-24                               837 ± 0%       837 ± 0%     ~     (all equal)
RESTAuthzAllow10Paths-24                               821 ± 0%       821 ± 0%     ~     (all equal)
RESTAuthzAllow100Paths-24                            4.34k ± 0%     4.34k ± 0%     ~     (all equal)
RESTAuthzAllow1000Paths-24                           40.4k ± 0%     40.4k ± 0%     ~     (p=1.000 n=10+10)
pkg:github.com/open-policy-agent/opa/test/e2e/logs/console goos:linux goarch:amd64
RESTConsoleDecisionLogger/1x0-24                       833 ± 0%       699 ± 0%  -16.09%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/10x0-24                      833 ± 0%       699 ± 0%  -16.09%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/100x0-24                     833 ± 0%       699 ± 0%  -16.09%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/1000x0-24                    832 ± 0%       698 ± 0%  -16.11%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/1x1-24                       884 ± 0%       755 ± 0%  -14.59%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/10x1-24                      884 ± 0%       755 ± 0%  -14.59%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/100x1-24                     884 ± 0%       755 ± 0%  -14.62%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/1000x1-24                    883 ± 0%       754 ± 0%  -14.61%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/10x10-24                     981 ± 0%       851 ± 0%  -13.22%  (p=0.000 n=9+10)
RESTConsoleDecisionLogger/100x10-24                    981 ± 0%       851 ± 0%  -13.25%  (p=0.000 n=8+8)
RESTConsoleDecisionLogger/1000x10-24                   979 ± 0%       850 ± 0%  -13.18%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/100x100-24                 1.90k ± 0%     1.77k ± 0%   -6.80%  (p=0.002 n=8+10)
RESTConsoleDecisionLogger/1000x100-24                1.89k ± 0%     1.76k ± 0%   -6.81%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/1000x1000-24               10.9k ± 0%     10.8k ± 0%   -1.17%  (p=0.000 n=10+10)
pkg:github.com/open-policy-agent/opa/test/e2e/logs/remote goos:linux goarch:amd64
RESTDecisionLogAuthzForbidAuthn-24                     737 ± 0%       605 ± 0%  -17.95%  (p=0.000 n=10+9)
RESTDecisionLogAuthzForbidPath-24                    1.16k ± 0%     1.02k ± 0%  -11.47%  (p=0.000 n=10+9)
RESTDecisionLogAuthzForbidMethod-24                  1.16k ± 0%     1.03k ± 0%  -11.34%  (p=0.000 n=10+10)
RESTDecisionLogAuthzAllow10Paths-24                  1.15k ± 0%     1.02k ± 0%  -11.50%  (p=0.000 n=9+10)
RESTDecisionLogAuthzAllow100Paths-24                 4.67k ± 0%     4.54k ± 0%   -2.83%  (p=0.000 n=10+10)
RESTDecisionLogAuthzAllow1000Paths-24                40.7k ± 0%     40.5k ± 0%   -0.32%  (p=0.000 n=10+9)
RESTRemoteDecisionLogger/1x0-24                        672 ± 0%       538 ± 0%  -19.94%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/10x0-24                       672 ± 0%       538 ± 0%  -19.94%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/100x0-24                      672 ± 0%       538 ± 0%  -19.94%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/1000x0-24                     672 ± 0%       538 ± 0%  -19.94%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/1x1-24                        717 ± 0%       588 ± 0%  -17.99%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/10x1-24                       717 ± 0%       588 ± 0%  -17.99%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/100x1-24                      717 ± 0%       588 ± 0%  -17.99%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/1000x1-24                     717 ± 0%       588 ± 0%  -17.99%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/10x10-24                      813 ± 0%       684 ± 0%  -15.87%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/100x10-24                     813 ± 0%       684 ± 0%  -15.87%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/1000x10-24                    813 ± 0%       684 ± 0%  -15.87%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/100x100-24                  1.73k ± 0%     1.60k ± 0%   -7.48%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/1000x100-24                 1.73k ± 0%     1.60k ± 0%   -7.48%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/1000x1000-24                10.7k ± 0%     10.6k ± 0%   -1.20%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1x0-24           1.05k ± 0%     0.92k ± 0%  -12.77%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/10x0-24          1.05k ± 0%     0.92k ± 0%  -12.77%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/100x0-24         1.05k ± 0%     0.92k ± 0%  -12.77%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1000x0-24        1.05k ± 0%     0.92k ± 0%  -12.77%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1x1-24           1.09k ± 0%     0.96k ± 0%  -11.79%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/10x1-24          1.09k ± 0%     0.96k ± 0%  -11.79%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/100x1-24         1.09k ± 0%     0.96k ± 0%  -11.79%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1000x1-24        1.09k ± 0%     0.96k ± 0%  -11.79%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/10x10-24         1.19k ± 0%     1.06k ± 0%  -10.84%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/100x10-24        1.19k ± 0%     1.06k ± 0%  -10.84%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1000x10-24       1.19k ± 0%     1.06k ± 0%  -10.84%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/100x100-24       2.10k ± 0%     1.97k ± 0%   -6.14%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1000x100-24      2.10k ± 0%     1.97k ± 0%   -6.14%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1000x1000-24     11.1k ± 0%     11.0k ± 0%   -1.16%  (p=0.000 n=10+10)
```

Note that in the future as we add more stuff into the decision log output these changes will help it scale better.